test: :white_check_mark: fix datacollector collector to str, refactoring tests and passing them
